Refactor: Store, Menu, Order 모듈 책임 분리 및 검증 로직 개선

+++ src/main/java/com/example/delivery/domain/menu/service/MenuServiceImpl.java
@@ -1,5 +1,7 @@
 package com.example.delivery.domain.menu.service;
 
+import com.example.delivery.common.exception.CustomException;
+import com.example.delivery.common.exception.enums.ErrorCode;
 import com.example.delivery.domain.store.entity.Store;
 import com.example.delivery.domain.store.repository.StoreRepository;
 import com.example.delivery.domain.menu.dto.RequestDto;
@@ -19,14 +21,15 @@ public class MenuServiceImpl implements MenuService {
     private final MenuRepository menuRepository;
     private final StoreRepository storeRepository;
 
+
     @Override
     @Transactional
-    public ResponseDto create(Long loginUserId, RequestDto requestDto) {
+    public ResponseDto create(Long ownerId, RequestDto requestDto) {
+        // 1. 가게 조회
+        Store store = storeRepository.findByIdOrElseThrow(requestDto.getId());
 
-//        Store store = storeRepository.findById(requestDto.getId())
-//                .orElseThrow(() -> new NotFoundException(ErrorCode.STORE_NOT_FOUND));
-        Store store = storeRepository.findMyStoreOrElseThrow(requestDto.getId(), loginUserId);
-        // 메뉴 생성은 사장님만 가능하게 해야되는데 사장님 검증 로직이 빠짐
+        // 2. 가게 오너 검증
+        store.validateOwner(ownerId);
 
         Menu menu = Menu.of(store, requestDto.getName(), requestDto.getPrice(), requestDto.getDescription());
         menuRepository.save(menu);
@@ -36,12 +39,13 @@ public class MenuServiceImpl implements MenuService {
 
     @Override
     @Transactional
-    public ResponseDto update(Long loginUserId, Long menuId, RequestDto requestDto) {
-        // 사장님인지 검증 로직
-        // 메뉴가 속한 가게의 소유자가 로그인된 유저인걸 검증하는 메서드
-        storeRepository.findMyStoreOrElseThrow(requestDto.getId(), loginUserId);
-        //메뉴 조회
-        Menu menu = menuRepository.findByIdOrElseThrow(loginUserId);
+    public ResponseDto update(Long ownerId, Long menuId, RequestDto requestDto) {
+        // 1. 메뉴 조회
+        Menu menu = menuRepository.findByIdOrElseThrow(ownerId);
+        // 2. 스토어 오너 검증
+        Store store = menu.getStore();
+        store.validateOwner(ownerId);
+
         // 메뉴 업데이트
         menu.update(requestDto.getName(), requestDto.getPrice(), requestDto.getDescription());
 
@@ -50,17 +54,17 @@ public class MenuServiceImpl implements MenuService {
 
     @Override
     @Transactional
-    public void delete(Long loginUserId, Long menuId) {
-        // 메뉴 조회
-        Menu menu = menuRepository.findByIdOrElseThrow(loginUserId);
-        // 사장인지 검증
-        storeRepository.findMyStoreOrElseThrow(menu.getStore().getId(), loginUserId);
+    public void delete(Long ownerId, Long menuId) {
+        Menu menu = menuRepository.findByIdOrElseThrow(ownerId);
+        Store store = menu.getStore();
+        store.validateOwner(ownerId);
         menu.delete();
     }
+
     @Override
     @Transactional(readOnly = true)
     public List<ResponseDto> getMenusByStore(Store store) {
-        return menuRepository.findActiveMenusByStore(store).stream()
+        return menuRepository.findAllByStoreAndDeletedFalse(store).stream()
                 .map(ResponseDto::from)
                 .collect(Collectors.toList());
